Moved client / server generators around
